+++ test/jest/__mock__/stripesSmartComponent.mock.js
@@ -1,8 +1,9 @@
 jest.mock('@folio/stripes/smart-components', () => ({
   ...jest.requireActual('@folio/stripes/smart-components'),
-  ConfigManager: jest.fn(({ lastMenu }) => (
+  ConfigManager: jest.fn(({ children, lastMenu }) => (
     <div>
       ConfigManager
+      {children}
       {lastMenu}
     </div>
   )),